# backend/src/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ReqAI – AI Requirement Analyzer API
  version: 1.0.0
  description: |
    Enterprise REST API for ReqAI — AI-powered requirement analysis platform.

    ## Authentication
    All protected endpoints require a Bearer JWT in the `Authorization` header.
    Obtain tokens via `POST /auth/login`. Refresh via `POST /auth/refresh` (httpOnly cookie).

    ## Response Envelope
    All responses follow a consistent envelope:
    ```json
    { "success": true, "data": {}, "meta": { "page": 1, "limit": 20, "total": 142 } }
    ```
    Errors:
    ```json
    { "success": false, "error": { "code": "VALIDATION_ERROR", "message": "...", "details": [] } }
    ```

  contact:
    name: ReqAI Engineering
    email: engineering@reqai.dev
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.reqai.dev/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (15-minute expiry). Obtain via POST /auth/login.

  # ─── Reusable Parameters ──────────────────────────────────────────────────────
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RequirementId:
      name: requirementId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  # ─── Reusable Responses ──────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
    Forbidden:
      description: Authenticated but lacks required permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Request body failed schema validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'

  # ─── Schemas ─────────────────────────────────────────────────────────────────
  schemas:
...
    RequirementStatus:
      type: string
      enum: [DRAFT, IN_ANALYSIS, ANALYZED, REVIEWED, APPROVED, ARCHIVED]
    RequirementType:
      type: string
      enum: [FUNCTIONAL, NON_FUNCTIONAL, BUSINESS, TECHNICAL, CONSTRAINT, ASSUMPTION]
    PriorityLevel:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:    { type: string, example: NOT_FOUND }
            message: { type: string, example: Requirement not found }
    ValidationErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:    { type: string, example: VALIDATION_ERROR }
            message: { type: string, example: Request validation failed }
            details:
              type: array
              items:
                type: object
                properties:
                  field:   { type: string, example: email }
                  message: { type: string, example: Invalid email address }
    PaginationMeta:
      type: object
      properties:
        page:       { type: integer, example: 1 }
        limit:      { type: integer, example: 20 }
        total:      { type: integer, example: 142 }
        totalPages: { type: integer, example: 8 }
        hasNext:    { type: boolean, example: true }
        hasPrev:    { type: boolean, example: false }

    # ── Auth Schemas ──────────────────────────────────────────────────────────
...
    CreateRequirementRequest:
      type: object
      required: [title, body]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 500
          example: User Authentication via OAuth 2.0
        body:
          type: string
          minLength: 10
          maxLength: 50000
          example: The system must allow users to authenticate using OAuth 2.0...
        type:
          $ref: '#/components/schemas/RequirementType'
        priority:
          $ref: '#/components/schemas/PriorityLevel'
        source:
          type: string
          example: Jira-PROJ-42
          nullable: true
        tags:
          type: array
          items: { type: string }
          example: [auth, oauth, security]
        parentId:
          type: string
          format: uuid
          nullable: true
    Requirement:
      type: object
      properties:
        id:            { type: string, format: uuid }
        projectId:     { type: string, format: uuid }
        title:         { type: string }
        body:          { type: string }
        type:          { $ref: '#/components/schemas/RequirementType' }
        priority:      { $ref: '#/components/schemas/PriorityLevel' }
        status:        { $ref: '#/components/schemas/RequirementStatus' }
        tags:          { type: array, items: { type: string } }
        version:       { type: integer, example: 3 }
        wordCount:     { type: integer, example: 128 }
        source:        { type: string, nullable: true }
        sourceFileUrl: { type: string, nullable: true }
        parentId:      { type: string, format: uuid, nullable: true }
        createdBy:     { type: string, format: uuid }
        analyzedAt:    { type: string, format: date-time, nullable: true }
        approvedAt:    { type: string, format: date-time, nullable: true }
        createdAt:     { type: string, format: date-time }
        updatedAt:     { type: string, format: date-time }
    RequirementListResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: array
          items: { $ref: '#/components/schemas/Requirement' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }

    # ── Analysis Schemas ──────────────────────────────────────────────────────
...
paths:
...
  /projects/{projectId}/requirements:
    parameters:
      - $ref: '#/components/parameters/ProjectId'

    get:
      tags: [Requirements]
      summary: List requirements for a project (paginated, filterable)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/RequirementStatus' }
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/RequirementType' }
        - name: priority
          in: query
          schema: { $ref: '#/components/schemas/PriorityLevel' }
        - name: search
          in: query
          schema: { type: string, maxLength: 300 }
          description: Full-text search against title and body
        - name: tags
          in: query
          schema: { type: string }
          description: Comma-separated tag filter (e.g. auth,oauth)
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, updatedAt, title, priority, status]
            default: createdAt
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated requirement list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RequirementListResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [Requirements]
      summary: Create a new requirement
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRequirementRequest' }
      responses:
        "201":
          description: Requirement created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/Requirement' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "422": { $ref: '#/components/responses/ValidationError' }
  /projects/{projectId}/requirements/upload:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [Requirements]
      summary: Upload a requirement document (.txt, .md, .pdf — max 10 MB)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                  description: Optional override for the requirement title
      responses:
        "201":
          description: Requirement created from uploaded file
        "400":
          description: Unsupported file type or empty file
        "401": { $ref: '#/components/responses/Unauthorized' }
  /projects/{projectId}/requirements/{requirementId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/RequirementId'

    get:
      tags: [Requirements]
      summary: Get a requirement by ID
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Requirement detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/Requirement' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }

    put:
      tags: [Requirements]
      summary: Update a requirement (auto-increments version if body changes)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRequirementRequest' }
      responses:
        "200":
          description: Requirement updated
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Requirements]
      summary: Soft-delete a requirement
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Requirement archived
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /projects/{projectId}/requirements/{requirementId}/versions:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/RequirementId'
    get:
      tags: [Requirements]
      summary: Get full version history of a requirement
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of version snapshots, newest first
  /projects/{projectId}/requirements/{requirementId}/links:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/RequirementId'
    post:
      tags: [Requirements]
      summary: Create a typed link between two requirements
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetId, linkType]
              properties:
                targetId:
                  type: string
                  format: uuid
                linkType:
                  type: string
                  enum: [RELATED, DEPENDS_ON, CONFLICTS_WITH, DUPLICATES, IMPLEMENTS]
      responses:
        "200":
          description: Link created

  # ── Analysis ────────────────────────────────────────────────────────────
...